ExceptionRecordAttacher
