+++ src/test/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/AttachToCaseCallbackServiceTest.java
@@ -7,7 +7,6 @@ import io.vavr.control.Validation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.client.model.request.DocumentType;
@@ -15,27 +14,20 @@ import uk.gov.hmcts.reform.bulkscan.orchestrator.client.model.request.DocumentUr
 import uk.gov.hmcts.reform.bulkscan.orchestrator.client.model.request.ScannedDocument;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.config.ServiceConfigItem;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.model.internal.ExceptionRecord;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.AttachScannedDocumentsValidator;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.ExceptionRecordValidator;
-import uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.callback.PaymentsHelper;
 import uk.gov.hmcts.reform.bulkscan.orchestrator.services.config.ServiceConfigProvider;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 
 import java.util.Map;
-import java.util.Optional;
 
 import static java.time.LocalDateTime.now;
-import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
-import static java.util.Collections.emptyMap;
 import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoInteractions;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CallbackValidationsTest.JOURNEY_CLASSIFICATION;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.CONTAINS_PAYMENTS;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.definition.ExceptionRecordFields.ENVELOPE_ID;
@@ -54,22 +46,13 @@ class AttachToCaseCallbackServiceTest {
     private ServiceConfigProvider serviceConfigProvider;
 
     @Mock
-    private CcdApi ccdApi;
-
-    @Mock
     private ExceptionRecordValidator exceptionRecordValidator;
 
     @Mock
     private ExceptionRecordFinalizer exceptionRecordFinalizer;
 
     @Mock
-    private CcdCaseUpdater ccdCaseUpdater;
-
-    @Mock
-    private PaymentsProcessor paymentsProcessor;
-
-    @Mock
-    private AttachScannedDocumentsValidator scannedDocumentsValidator;
+    private ExceptionRecordAttacher exceptionRecordAttacher;
 
     private static final String BULKSCAN_ENVELOPE_ID = "some-envelope-id";
     private static final String JURISDICTION = "BULKSCAN";
@@ -95,7 +78,7 @@ class AttachToCaseCallbackServiceTest {
         .data(ImmutableMap.<String, Object>builder()
             .put(JOURNEY_CLASSIFICATION, SUPPLEMENTARY_EVIDENCE_WITH_OCR.name())
             .put(SEARCH_CASE_REFERENCE, EXISTING_CASE_ID)
-            .put(OCR_DATA, asList(ImmutableMap.of("firstName", "John")))
+            .put(OCR_DATA, singletonList(ImmutableMap.of("firstName", "John")))
             .put(SCANNED_DOCUMENTS, ImmutableList.of(EXISTING_DOC))
             .put(CONTAINS_PAYMENTS, YES)
             .put(ENVELOPE_ID, BULKSCAN_ENVELOPE_ID)
@@ -103,28 +86,18 @@ class AttachToCaseCallbackServiceTest {
         )
         .build();
 
-    private static final CaseDetails EXISTING_CASE_DETAILS = CaseDetails.builder()
-        .jurisdiction(JURISDICTION)
-        .caseTypeId(EXISTING_CASE_TYPE)
-        .id(Long.parseLong(EXISTING_CASE_ID))
-        .data(emptyMap())
-        .build();
-
     @BeforeEach
     void setUp() {
         attachToCaseCallbackService = new AttachToCaseCallbackService(
             serviceConfigProvider,
-            ccdApi,
             exceptionRecordValidator,
             exceptionRecordFinalizer,
-            ccdCaseUpdater,
-            paymentsProcessor,
-            scannedDocumentsValidator
+            exceptionRecordAttacher
         );
 
         exceptionRecord = getExceptionRecord();
         given(exceptionRecordValidator.getValidation(CASE_DETAILS)).willReturn(Validation.valid(exceptionRecord));
-        given(ccdApi.getCase(anyString(), anyString())).willReturn(EXISTING_CASE_DETAILS);
+        //given(ccdApi.getCase(anyString(), anyString())).willReturn(EXISTING_CASE_DETAILS);
         configItem = new ServiceConfigItem();
         configItem.setUpdateUrl("url");
         configItem.setService(SERVICE_NAME);
@@ -135,9 +108,12 @@ class AttachToCaseCallbackServiceTest {
     void process_should_process_supplementary_evidence_with_ocr() {
         // given
         given(exceptionRecordValidator.mandatoryPrerequisites(any())).willReturn(Validation.valid(null));
-        given(ccdCaseUpdater.updateCase(
-            exceptionRecord, SERVICE_NAME, true, IDAM_TOKEN, USER_ID, EXISTING_CASE_ID, EXISTING_CASE_TYPE
-        )).willReturn(Optional.empty());
+        given(exceptionRecordAttacher.tryAttachToCase(
+            any(AttachToCaseEventData.class),
+            any(CaseDetails.class),
+            anyBoolean()
+        ))
+            .willReturn(Either.right(EXISTING_CASE_ID));
 
         // when
         Either<ErrorsAndWarnings, Map<String, Object>> res = attachToCaseCallbackService.process(
@@ -150,17 +126,6 @@ class AttachToCaseCallbackServiceTest {
 
         // then
         assertThat(res.isRight()).isTrue();
-        verify(ccdCaseUpdater)
-            .updateCase(exceptionRecord, SERVICE_NAME, true, IDAM_TOKEN, USER_ID, EXISTING_CASE_ID, EXISTING_CASE_TYPE);
-
-        // and
-        var paymentsDataCaptor = ArgumentCaptor.forClass(PaymentsHelper.class);
-        verify(paymentsProcessor)
-            .updatePayments(paymentsDataCaptor.capture(), eq(CASE_REF), eq(JURISDICTION), eq(EXISTING_CASE_ID));
-        assertThat(paymentsDataCaptor.getValue()).satisfies(data -> {
-            assertThat(data.containsPayments).isEqualTo(CASE_DETAILS.getData().get(CONTAINS_PAYMENTS).equals(YES));
-            assertThat(data.envelopeId).isEqualTo(BULKSCAN_ENVELOPE_ID);
-        });
 
         // and exception record should be finalized
         verify(exceptionRecordFinalizer).finalizeExceptionRecord(
@@ -174,11 +139,12 @@ class AttachToCaseCallbackServiceTest {
     void process_should_not_update_case_if_error_occurs() {
         // given
         given(exceptionRecordValidator.mandatoryPrerequisites(any())).willReturn(Validation.valid(null));
-        given(ccdCaseUpdater.updateCase(
-            exceptionRecord, SERVICE_NAME, true, IDAM_TOKEN, USER_ID, EXISTING_CASE_ID, EXISTING_CASE_TYPE
-        )).willReturn(
-            Optional.of(new ErrorsAndWarnings(asList("error1"), asList("warning1")))
-        );
+        given(exceptionRecordAttacher.tryAttachToCase(
+            any(AttachToCaseEventData.class),
+            any(CaseDetails.class),
+            anyBoolean())
+        )
+            .willReturn(Either.left(new ErrorsAndWarnings(singletonList("error1"), singletonList("warning1"))));
 
         // when
         Either<ErrorsAndWarnings, Map<String, Object>> res = attachToCaseCallbackService.process(
@@ -191,12 +157,8 @@ class AttachToCaseCallbackServiceTest {
 
         // then
         assertThat(res.isLeft()).isTrue();
-        assertThat(res.getLeft().getWarnings()).isEqualTo(asList("warning1"));
-        assertThat(res.getLeft().getErrors()).isEqualTo(asList("error1"));
-
-        verify(ccdCaseUpdater)
-            .updateCase(exceptionRecord, SERVICE_NAME, true, IDAM_TOKEN, USER_ID, EXISTING_CASE_ID, EXISTING_CASE_TYPE);
-        verifyNoInteractions(paymentsProcessor);
+        assertThat(res.getLeft().getWarnings()).isEqualTo(singletonList("warning1"));
+        assertThat(res.getLeft().getErrors()).isEqualTo(singletonList("error1"));
     }
 
     private static Map<String, Object> document(String filename, String documentNumber) {
